journal/requirements.txt
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.2.2
numpy==1.19.2
pandas==0.24.2
robin-stocks==1.5.1
graphene-django==2.13.0
